Moved rules selector
