+++ src/containers/AppBody.jsx
@@ -7,8 +7,8 @@ import { setState } from '../actions/app';
 import AnnotationDisplay from './AnnotationDisplay';
 import SnippetArea from './SnippetArea';
 import Title from './Title';
+import RulesSelector from './RulesSelector';
 import Error from '../components/Error';
-import RulesSelector from '../components/RulesSelector';
 import { getFirstAnnotation } from '../util/annotations';
 
 const API_URL = process.env.API_URL;
