+++ src/containers/RulesSelector.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
 
-import RuleLabel from './RuleLabel';
+import RuleLabel from '../components/RuleLabel';
 import { toggleRule } from '../actions/app';
 
 const styles = {
